Update patch set 1
Patch Set 1:

> I'm happy to rebase if you think it's a useful change. If someone feels that's unlikely, please let me know and I'll abandon.

OK, I hadn't really looked at the CL.  That's a good question.  The CL description doesn't fully describe the change, but it looks like you are inserting a check in any conversion of uintptr to unsafe.Pointer that the value being converted is >= 4096.  In a sense you are changing the language, in that unsafe.Pointer(uintptr(1)) compiles today but will fail to compile with your change, even if the code is never executed; that is problematic.  This is also likely to break some existing code that uses cgo to call C functions that take void* arguments but are passed integer values; such code is dubious and perhaps broken today but it may sometimes work, and with this change it would definitely break.  And, of course, you are only detecting one class of invalid pointers.

I think there is a reasonable argument that in a language that purports to be memory safe we should have a way to verify that created pointers are valid, and checking the conversion from integer to unsafe.Pointer is clearly one reasonable place to check that.  But it's really a change to the existing language, and the place to hash that out is not in CL comments.

If you are still interested in this, I recommend that you open a proposal (seehttps://github.com/golang/proposal/blob/master/README.md) and we can see what happens there.

Thanks.

Patch-set: 1