cmd/compile: attempt to deflake debug_test.go
Removed timeouts from normal path and ensure that they were
not needed and that reference files did not change.

By default, testing with gdb uses compilation with
inlining disabled to prevent dependence on library code
(it's a bug that library code is seen while Nexting, but
the bug is current behavior).

By default, testing with gdb ignores repeated lines that
occur when nexting.  This appears to sometimes be timing-
dependent and is the observed source of flakiness in testing
so far.

Added flags -i and -r to make gdb testing compile with
inlining and be sensitive to repeats in the next stream.
This is for developer-testing and so we can describe these
problems in bug reports.

Removed testing.Short() exclusion to allow flake-testing
in vivo (need to figure out how to obtain multiple TryBot
runs.)

Fourth Trybot run.

Updates #22206.

Change-Id: I9a30ebbc65aa0153fe77b1858cf19743bdc985e4