cmd/compile: attempt to deflake debug_test.go
DO NOT REVIEW -- test is wrapped in a loop for flake testing

Removed timeouts from normal path and ensure that they were
not needed and that reference files did not change.

Use "tbreak" instead of "break" with gdb to reduce chance
of multiple hits on main.main.

By default, testing ignores repeated lines that occur when
nexting.  This appears to sometimes be timing-dependent and
is the observed source of flakiness in testing so far.
Note that these can also be signs of a bug in the generated
debugging output, but it is one of the less-confusing bugs
that can occur.

By default, testing with gdb uses compilation with
inlining disabled to prevent dependence on library code
(it's a bug that library code is seen while Nexting, but
the bug is current behavior).

Exclude src/runtime/proc.go from the next history so that
the code between return from main and process exit (in the
runtime) is not sensitive to changes in line numbers.

Added flags -i and -r to make gdb testing compile with
inlining and be sensitive to repeats in the next stream.
This is for developer-testing and so we can describe these
problems in bug reports.

Removed testing.Short() exclusion to allow flake-testing
in vivo (need to figure out how to obtain multiple TryBot
runs.)

Fifth Trybot run.

Updates #22206.

Change-Id: I9a30ebbc65aa0153fe77b1858cf19743bdc985e4