Update patch set 1
Patch Set 1:

> I don't agree that this changes the language, since this constraint already exists (but in a subtle way). The noted code compiles, but could cause a runtime panic. Anyways, I don't feel strongly either way about this change moving forward, but I knocked out a proposal here [1] to collect feedback. Hopefully that explains it well.

When there is code that used to compile, and the compiler is changed so that that code no longer compiles, that is a language change.  Adding a new runtime panic is quite different: it means that the code only fails if it is executed; that is not a language change.  It's perfectly reasonable to write code that is sure to fail, provided it is never executed; for example, this happens routinely when using build tags for portability.

Thanks for writing the proposal.

Patch-set: 1