runtime: work around android failure
For reasons unknown, on Android, the gsignal stack appears to be 12K
and C heap allocated (or maybe directly mmap'd). I have no idea why,
but it's breaking the build right now because we try to stackfree it.

Work around this and fix the build by only doing the stackfree if it's
Go-allocated. This isn't the right fix, but it'll do for now.

Change-Id: Ic9b7b9339a561a321b125b715a0cbff7a222f3e1