Update patch set 1
Patch Set 1:

The original motivation was to change the gofmt command to ignore the vendor package.

> It is not a subset, it is a wrapper. The important difference between go fmt and gofmt is the list of packages passed to the binary by go fmt, a list that for example excludes vendored packages.

However, as pointed out above, `go fmt` is the gofmt's wrapper
 (https://github.com/golang/go/issues/22173#issuecomment-334975082).
It means `go fmt` has the responsibility to collect the target go files and pass them to the gofmt command. `go fmt` cares (ignores) vendored packages, so we need to use `go fmt` to format code with ignoring the vendored package.
However `go fmt` cannot access the full features of gofmt because the flags that `go fmt` available are limited (especially the `-s` flag as mentiond on the issue). I feel it is inconvenient.

The points are as follows.

Now:
- `go fmt` ignores vendored packaged
- `go fmt -s` doesn’t work (`-s` is not supported)
- `gofmt -s` works, but doesn’t ignore vendored packages

So:
- I want to use `go fmt -s`
  - That simplifies code and ignores vendored packages
- And I want to treat other options as well

> I think the only extra flag we want to expose is -s which is what people usually want.

This is a correct indication, I have no intention to object. However I don't know much whether it can satisfy the user's use case. Then I think that it is more convenient to expose all functions to users (similar opinion: https://github.com/golang/go/issues/22173#issuecomment-334975082). Probably users select the function by himself.

> For one, we probably should remove the -r option from gofmt proper and make it a separate tool

It is beyond the scope of this topic, I think.

Patch-set: 1
Reviewer: Gerrit User 25079 <25079@62eb7196-b449-3ce5-99f1-c037f21e1705>