Update patch set 7
Patch Set 7:

This looks more reasonable but I would still leave it disabled in short mode. The report at #22206 says it ran for 10 seconds before failing, which is far too long for the all.bash runs, flaky or not. It's getting to be time for another all.bash cleanup as it is.

Patch-set: 7
Reviewer: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>