Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Change has been successfully cherry-picked as 44d9e96da9b7625be81f2c7eacf73fcc609874ce by Austin Clements
> 
> It seems this CL fixed android but broke ios instead. Sorry.
> 
> https://build.golang.org/log/9eb4cbd4c3827851ec2a37c2046f69cf1da94c84
> https://build.golang.org/log/dc0dfc95245ceef416ec2b722182ee907c772f8a

Sigh.

Theory: It looks like pthread_kill might be using SIGINT internally, so it doesn't matter that we've blocked signals when thread exiting, it's going to call our signal handler anyway. But, if we're on an OS-allocated signal stack, then at that point we've cleared gsignal.stack to indicate that we don't own it (new in this CL). The cleared gsignal.stack then breaks sigtramp when we get the self-SIGINT.

If that's right, then unminitSignals shouldn't clear gsignal.stack and we should use some other indicator (like newSigstack) to avoid freeing an OS signal stack.

Patch-set: 2