Update patch set 1
Patch Set 1:

> Is this intended as a temporary fix? I'd rather not simply plaster over the problem.

Yes. I haven't figured out what is the issue yet.
My intent was to bring the builders back to "ok" in the meantime, so I'll be able to catch issues in the following commits.

> Do you know where the nil pointer dereference that's causing startProcess to crash is (e.g., https://build.golang.org/log/c14eefc1c70d14ad2d0f768ff9f13222fa8d435c)? It must have something to do with the goroutine that startProcess starts. One would presume it's in the LockOSThread call under startTemplateThread, but what I find confusing is that the crash is happening on the system stack and we never obviously switch to the system stack there.

I don't know where the nil pointer dereference happens yet, especially since the PC address is invalid. It doesn't seem to happen in the LockOSThread function. Depending on the crash, it happens either on read or on write. Do you know a way to find it?

Patch-set: 1
Reviewer: Gerrit User 5060 <5060@62eb7196-b449-3ce5-99f1-c037f21e1705>