runtime/cgo: return from crosscall_arm1
crosscall_arm1 is missing a "ret" instruction at the end. This never
mattered before, but now that the runtime can exit threads, it's
possible to return to this function and it's important that it return
back to pthreads in order to exit the thread.

Somehow this hasn't been crashing all of the */arm cgo tests. I don't
know why not.

Updates #22248.

Change-Id: I4e5ed0c0226241779d1000b15b43056b7ff0e260