Update patch set 1
Patch Set 1:

Thanks for submitting your first change Robert. Usually, any changes that are made to the language or to the code are initially discussed on the mailing list or on the GitHub issues. I can not find the discussion relating to this change. Please read https://golang.org/doc/contribute.html#Design

If there already is a discussion and I've missed it, then my apologies. Otherwise, please considering having it before committing to any actual code.

Patch-set: 1
CC: Gerrit User 5465 <5465@62eb7196-b449-3ce5-99f1-c037f21e1705>