Update patch set 1
Patch Set 1:

> But I don't think anybody would use `go fmt` without -l and -w.  It would be very hard to interpret the output.
> 
> I tend to agree that we should expose -s and call it a day.

How about `-dryrun` option? This option is an opposite of `-w`.
I think that there is a use case to obtain a list of files to be formatted without side effects (as a matter of fact, I want it).

Patch-set: 1
Reviewer: Gerrit User 25079 <25079@62eb7196-b449-3ce5-99f1-c037f21e1705>