Update patch set 1
Patch Set 1:

> > I tend to agree that we should expose -s and call it a day.
> 
> How about `-dryrun` option? This option is an opposite of `-w`.
> I think that there is a use case to obtain a list of files to be formatted without side effects (as a matter of fact, I want it).

The purpose of this CL is getting blurrier and a bit out of control. -dryrun isn't part of gofmt, so it doesn't follow that it should be part of go fmt. If you'd want such an option, you should raise an issue in the tracker about it.

Patch-set: 1