Update patch set 2
Patch Set 2:

(1 comment)

It seems to me if you're going to add examples for Printf it would be nice to use a format that makes Printf worth doing, perhaps a floating-point value neatly formatted with decimals and so on. As written it's not a very instructive example you're adding.

Patch-set: 2
Reviewer: Gerrit User 5015 <5015@62eb7196-b449-3ce5-99f1-c037f21e1705>