Update patch set 10
Patch Set 10:

I'm not sure if this is the right place to do this but I believe this change broke some of my code.  I just made an update to a library of mine that I have building on Travis CI against go tip and it just failed with the following error:

2017/10/17 18:03:48 comments are out of order. /home/travis/gopath/src/github.com/aphistic/gomol/attrs.go:147:1 was before -.

The file it's mentioning hasn't changed in 2 months and I can't decipher what the error actually means in this context.  The build itself is at https://travis-ci.org/aphistic/gomol/jobs/289149613#L544. The code in question is at https://github.com/aphistic/gomol/blob/fallback-logger/attrs.go#L147.

Patch-set: 10
CC: Gerrit User 25216 <25216@62eb7196-b449-3ce5-99f1-c037f21e1705>