Update patch set 1
Patch Set 1:

Thanks for the quick fix.

How do you feel about leaving this out of the upcoming point release? It does seem that this probably avoids some unexpected concurrent calls, but the flip side is that it may introduce deadlocks. Especially since we have no experience with the CL yet and because it's only fixing a hypothetical problem, leaving it out of the upcoming point release seems like the most prudent course of action. This CL seems to be risk without practical reward.

In contrast, users on #21117 seems to believe that CL 65731 fixes a real problem they are hitting very frequently, so the risk is outweighed by a demonstrable benefit, so I'd be inclined to take CL 65731.

Patch-set: 1
Reviewer: Gerrit User 5056 <5056@62eb7196-b449-3ce5-99f1-c037f21e1705>