Update patch set 1
Patch Set 1: Code-Review+1

(1 comment)

This looks good to me. I pulled the patch and tried compiling some stuff with it; the generated DWARF looked as expected.

One thing to think about is whether it makes sense to capture the value of important flags that don't actually appear on the command line, in case the default value for those flags changes at some point in the future or is different due to the configuration/GOOS/GOARCH. Gcc does this to a certain extent. For example:

% gcc -g himom.c ; objdump --dwarf=info a.out | fgrep AT_producer
<c>   DW_AT_producer    : (indirect string, offset: 0x2f): GNU C 4.8.4 -mtune=generic -march=x86-64 -g -fstack-protector

I didn't specify "-mtune=generic" or "-fstack-protector" on the command line, but gcc put them in anyhow, assumption being that it wants to capture the default setting that has been baked into the compiler via configure.

Thinking in those terms, would it make sense to always capture the "-l" value, in case at some point we switch the default from "-l=1" to "-l=2", or something like that? This is a somewhat contrived example, maybe it's not worth bothering about.

Patch-set: 1
Reviewer: Gerrit User 14020 <14020@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+1