Update patch set 5
Patch Set 5:

What seems to be happening is that for test binaries in package main we wind up building one archive containing all the usual source files and a second archive containing the generated ._testmain.go. Since the test itself is in package main, both of these archives are package main, so when we try to link them we get a duplicate symbol. For Linux, it's only happening on the noopt builder because that's the only place where there are compiler flags; otherwise we just don't wind up with the producer symbol. I'm not sure why it's happening on a few of the other builders, but I suspect it's something similar.

Patch-set: 5