Update patch set 5
Patch Set 5:

(1 comment)

We need a story around compatibility agreement. If I marshal the state to disk on Go1.10, and then try to unmarshal it years later with Go1.15, will it still work?

We should add version information to the binary output, and document what the expectation is. My personal opinion is that it should be possible to provide forward compatibility even if the underlying struct changes (but I'm no crypto expert). It's also possible that an optimization requires a different memory layout and the serialized form is not tranformable to the optimized version.

Patch-set: 5
CC: Gerrit User 9735 <9735@62eb7196-b449-3ce5-99f1-c037f21e1705>