Update patch set 5
Patch Set 5:

In order to avoid accidental mixing of encoding binary from one hash to another, we should add versioning and a magic prefix to every output.

Similar to the time package, we can prefix every output with a \x01 byte as the initial version.

https://github.com/golang/go/blob/b79337212585a473238a37421cefa9931b2fca4c/src/time/time.go#L1110

After that, we can append some type of format marker. One simple scheme is to encode the package name: "md5", "sha256", "adler32", etc... Another alternative is to just choose some byte that is unique in relation to all the other hash implementations. The advantage of the string approach is that manual inspection of the binary output clearly shows what it is. The downside is the extra space usage, which is most significant for short hashes like adler32.

Patch-set: 5