Update patch set 1
Patch Set 1:

> it seems strictly better (and not too hard) to close the connection if we see a Transfer-Encoding header in 1xx, 204, or 304 response.
>

That's a really interesting question that I had not considered. While I agree with the general philosophy of "if it's ambiguous, do whatever chrome does", not closing the connection just seems like it will be a problem in the future. I've opened up an issue on the http11bis issue tracker, which will hopefully provide clarity what clients are to do in this situation. If the answer is that TE header should not be allowed on a 204 or 304, then I'd like to see the change to close the connection happen in a follow up CL.

Patch-set: 1
Reviewer: Gerrit User 5673 <5673@62eb7196-b449-3ce5-99f1-c037f21e1705>