Update patch set 1
Patch Set 1:

(2 comments)

>Is it not possible to produce a real entry for these? I'd think that if a value >isn't SSAable, then it has to live on its stack location permanently, and we >could describe it as such.

For stacked locals it's possible that you could have two arrays "a [10]int" and "b [10]int" that are accessed in disjoint portions of the function, so the optimizer decides to assign them to the same stack location based on liveness. Without a location list, that would mean that if you types "print a" at some point where "a" is not live, GDB would happily print out the value of "b". That would be the case that I'd be concerned about.

With that said, you can make a good argument for putting in the stack location as opposed to the empty range list-- I think it would help the overall debugging experience. I am OK with either way.

Patch-set: 1
Reviewer: Gerrit User 14020 <14020@62eb7196-b449-3ce5-99f1-c037f21e1705>