Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> This is almost certainly going to break callers that assume req.TransferEncoding has been lower cased. We can't accept this change unless you can demonstrate that callers will not be broken.

The problem is current code is not correct, it is changing the incoming values of Transfer-Encoding - which is not correct so it is sort of manipulating the incoming data here which is incorrect.

Now the question of clients/servers depending on this should never depend on the case-sensitivity of the incoming values of r.TransferEncoding, neither its mentioned in the documentation that is what is expected of the callers to look for. In such a dependent scenario i would think that they are broken as well.

I don't see a way to make this backward compatible though unless we can keep the current fields as is and add another field which contains lets say r.TransferEncodingRaw values?

Patch-set: 1
Reviewer: Gerrit User 11900 <11900@62eb7196-b449-3ce5-99f1-c037f21e1705>