Update patch set 1
Patch Set 1: Code-Review-2

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > This is almost certainly going to break callers that assume req.TransferEncoding has been lower cased. We can't accept this change unless you can demonstrate that callers will not be broken.
> 
> The problem is current code is not correct, it is changing the incoming values of Transfer-Encoding - which is not correct so it is sort of manipulating the incoming data here which is incorrect.

Why is it not correct? The registered names are all lower-case and there is no language in the RFC that forbids what net/http is doing.
https://tools.ietf.org/html/rfc7230#section-4

I'd argue that it is more incorrect to rely on capitalization of transfer coding names, since coding name comparison is defined to be case insensitive.

> Now the question of clients/servers depending on this should never depend on the case-sensitivity of the incoming values of r.TransferEncoding, neither its mentioned in the documentation that is what is expected of the callers to look for. In such a dependent scenario i would think that they are broken as well.

I would consider that a documentation bug, not a code bug. Here's one piece of code I found that relies on the lower-case canonicalization. I only searched for a few seconds ... I'm sure there's a lot more code like this.
https://github.com/google/martian/blob/master/messageview/messageview.go#L88

> I don't see a way to make this backward compatible though unless we can keep the current fields as is and add another field which contains lets say r.TransferEncodingRaw values?

I have some sympathy for the opinion that net/http should not canonicalize any header field values as it currently does for TransferEncoding. However, that ship has sailed.

As for adding a new public field to http.Request, we try to avoid doing that unless there is good reason. I am not seeing a good reason in this case. It sounds like you're trying to fingerprint request headers at a low level. This is not something we support. e.g., you might as well be computing your fingerprint as a function of header field order, which is not preserved by net/http. If you actually need to fingerprint low-level things like capitalization of case-insenstive fields, I would tee the connection into a manually-written header parser.

Patch-set: 1
Label: Code-Review=-2